Move device_number and block_number to Link

+++ kernel/c/buf.h
@@ -1,7 +1,7 @@
 struct buf
 {
   uchar *data;
-  uint32 blockno;
+  uint64 blockno;
   uint64 cache_index;
   void *original;
 };
